+++ packages/typespec-go/CHANGELOG.md
@@ -5,6 +5,7 @@
 ### Features Added
 
 * Emit `LICENSE.txt` file (defaults to MIT license with a Microsoft Copyright).
+* Added support for `nextLinkVerb` for pageable operations.
 
 ## 0.8.3 (2025-10-14)